+++ install/02_configure_systemd.sh
@@ -17,14 +17,22 @@ echo "Installation de la configuration systemd-tmpfiles…"
 envsubst '${USER}' < "${BASE_DIR}/install/assets/tmpfiles-geonature.conf" | sudo tee /etc/tmpfiles.d/geonature.conf
 sudo systemd-tmpfiles --create /etc/tmpfiles.d/geonature.conf
 
-echo "Installation du service-file systemd…"
+echo "Installation des fichiers de service systemd…"
 envsubst '${USER} ${BASE_DIR}' < "${BASE_DIR}/install/assets/geonature.service" | sudo tee /etc/systemd/system/geonature.service
+envsubst '${USER} ${BASE_DIR}' < "${BASE_DIR}/install/assets/geonature-worker.service" | sudo tee /etc/systemd/system/geonature-worker.service
 sudo systemctl daemon-reload
 
 echo "Activation de geonature au démarrage…"
-sudo systemctl enable geonature || exit 1
+sudo systemctl enable geonature
+if [ "${WORKER}" == true ]; then
+    echo "Activation de geonature-worker au démarrage…"
+    sudo systemctl enable geonature-worker
+fi
 
 echo "Installation de la configuration logrotate…"
 envsubst '${USER}' < "${BASE_DIR}/install/assets/log_rotate" | sudo tee /etc/logrotate.d/geonature
 
 echo "Vous pouvez maintenant démarrer GeoNature avec la commande : sudo systemctl start geonature"
+if [ "${WORKER}" == true ]; then
+    echo "Vous pouvez maintenant démarrer le worker GeoNature avec la commande : sudo systemctl start geonature-worker"
+fi