+++ install/migration/migration.sh
@@ -190,5 +190,8 @@ if [ -f "/var/log/geonature.log" ]; then
 fi
 
 sudo systemctl restart geonature
+if [ "${WORKER}" = true ]; then
+    sudo systemctl restart geonature-worker
+fi
 
 deactivate